feat(oidc): address review comments
