+++ pkg/api/routes_internal_test.go
@@ -1,32 +1,14 @@
 package api
 
 import (
-	"errors"
-	"io"
-	"net/http"
-	"net/http/httptest"
 	"reflect"
 	"strings"
 	"testing"
 
 	"zotregistry.dev/zot/v2/pkg/api/config"
-	"zotregistry.dev/zot/v2/pkg/api/constants"
-	"zotregistry.dev/zot/v2/pkg/log"
 	"zotregistry.dev/zot/v2/pkg/storage"
 	storageTypes "zotregistry.dev/zot/v2/pkg/storage/types"
 )
-
-var errTestTokenProxyRead = errors.New("test token proxy read error")
-
-type errTokenProxyReadCloser struct{}
-
-func (errTokenProxyReadCloser) Read(_ []byte) (int, error) {
-	return 0, errTestTokenProxyRead
-}
-
-func (errTokenProxyReadCloser) Close() error {
-	return nil
-}
 
 func TestParseRangeHeader(t *testing.T) {
 	t.Parallel()
@@ -218,451 +200,5 @@ func TestIsBlobRedirectEnabled(t *testing.T) {
 	// Default storage remains disabled even when a specific subpath enables redirect.
 	if routeHandler.isBlobRedirectEnabled("b/repo") {
 		t.Fatal("expected redirect to be disabled for default storage")
-	}
-}
-
-func TestTokenProxyRequestBody(t *testing.T) {
-	t.Parallel()
-
-	t.Run("nil body", func(t *testing.T) {
-		t.Parallel()
-
-		body, contentLength, err := tokenProxyRequestBody(&http.Request{}, "upstream")
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		if body != nil {
-			t.Fatal("expected nil body")
-		}
-
-		if contentLength != 0 {
-			t.Fatalf("expected content length 0, got %d", contentLength)
-		}
-	})
-
-	t.Run("passes through non-form body", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, "/token", strings.NewReader("raw-body"))
-		request.Header.Set("Content-Type", "application/json")
-		request.ContentLength = int64(len("raw-body"))
-
-		body, contentLength, err := tokenProxyRequestBody(request, "upstream")
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		bodyBytes, err := io.ReadAll(body)
-		if err != nil {
-			t.Fatalf("failed to read body: %v", err)
-		}
-
-		if string(bodyBytes) != "raw-body" {
-			t.Fatalf("expected raw body to be preserved, got %q", string(bodyBytes))
-		}
-
-		if contentLength != int64(len("raw-body")) {
-			t.Fatalf("expected content length %d, got %d", len("raw-body"), contentLength)
-		}
-	})
-
-	t.Run("rewrites form service", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, "/token", strings.NewReader(
-			"grant_type=password&service=zot-service&scope=repository:test:pull",
-		))
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded; charset=utf-8")
-
-		body, _, err := tokenProxyRequestBody(request, "upstream")
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-
-		bodyBytes, err := io.ReadAll(body)
-		if err != nil {
-			t.Fatalf("failed to read body: %v", err)
-		}
-
-		got := string(bodyBytes)
-		if !strings.Contains(got, "service=upstream") {
-			t.Fatalf("expected service to be rewritten, got %q", got)
-		}
-
-		if !strings.Contains(got, "scope=repository%3Atest%3Apull") {
-			t.Fatalf("expected scope to be preserved, got %q", got)
-		}
-	})
-
-	t.Run("returns body read errors", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, "/token", nil)
-		request.Body = errTokenProxyReadCloser{}
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		_, _, err := tokenProxyRequestBody(request, "upstream")
-		if !errors.Is(err, errTestTokenProxyRead) {
-			t.Fatalf("expected read error, got %v", err)
-		}
-	})
-
-	t.Run("returns form parse errors", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, "/token", strings.NewReader("service=%zz"))
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		_, _, err := tokenProxyRequestBody(request, "upstream")
-		if !errors.Is(err, errInvalidTokenProxyForm) {
-			t.Fatalf("expected invalid form error, got %v", err)
-		}
-	})
-
-	t.Run("rejects oversized form body", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, "/token",
-			strings.NewReader(strings.Repeat("a", constants.MaxTokenRequestBodySize+1)))
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		_, _, err := tokenProxyRequestBody(request, "upstream")
-		if !errors.Is(err, errTokenRequestBodyTooLarge) {
-			t.Fatalf("expected body too large error, got %v", err)
-		}
-	})
-}
-
-func TestTokenProxyHeaders(t *testing.T) {
-	t.Parallel()
-
-	src := http.Header{}
-	src.Add("Connection", "X-Hop, X-Also-Hop")
-	src.Set("X-Hop", "drop")
-	src.Set("X-Also-Hop", "drop")
-	src.Set("Content-Length", "10")
-	src.Set("Keep-Alive", "timeout=5")
-	src.Set("X-Keep", "preserve")
-
-	dst := http.Header{}
-	copyTokenProxyHeaders(dst, src)
-
-	if dst.Get("X-Keep") != "preserve" {
-		t.Fatalf("expected X-Keep to be preserved, got %q", dst.Get("X-Keep"))
-	}
-
-	for _, header := range []string{"X-Hop", "X-Also-Hop", "Content-Length", "Keep-Alive", "Connection"} {
-		if dst.Get(header) != "" {
-			t.Fatalf("expected %s to be stripped, got %q", header, dst.Get(header))
-		}
-	}
-}
-
-func TestIsFormURLEncoded(t *testing.T) {
-	t.Parallel()
-
-	if isFormURLEncoded("application/%zz") {
-		t.Fatal("expected invalid content type to be rejected")
-	}
-}
-
-func TestProxyOIDCBearerTokenExchangeErrorsAndRedirect(t *testing.T) {
-	t.Parallel()
-
-	routeHandler := &RouteHandler{}
-
-	t.Run("rejects invalid proxy realm", func(t *testing.T) {
-		t.Parallel()
-
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			httptest.NewRecorder(),
-			httptest.NewRequest(http.MethodGet, "/token", nil),
-			&config.BearerConfig{ProxyRealm: "http://[::1", ProxyService: "upstream"},
-		)
-		if !errors.Is(err, errInvalidProxyRealm) {
-			t.Fatalf("expected invalid proxy realm error, got %v", err)
-		}
-	})
-
-	t.Run("rejects relative proxy realm", func(t *testing.T) {
-		t.Parallel()
-
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			httptest.NewRecorder(),
-			httptest.NewRequest(http.MethodGet, "/token", nil),
-			&config.BearerConfig{ProxyRealm: "/token", ProxyService: "upstream"},
-		)
-		if !errors.Is(err, errInvalidProxyRealm) {
-			t.Fatalf("expected invalid proxy realm error, got %v", err)
-		}
-	})
-
-	t.Run("rejects insecure proxy realm by default", func(t *testing.T) {
-		t.Parallel()
-
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			httptest.NewRecorder(),
-			httptest.NewRequest(http.MethodGet, "/token", nil),
-			&config.BearerConfig{ProxyRealm: "http://example.com/token", ProxyService: "upstream"},
-		)
-		if !errors.Is(err, errInvalidProxyRealm) {
-			t.Fatalf("expected invalid proxy realm error, got %v", err)
-		}
-	})
-
-	t.Run("returns form rewrite errors", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, "/token", strings.NewReader("service=%zz"))
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			httptest.NewRecorder(),
-			request,
-			&config.BearerConfig{
-				ProxyRealm:              "http://example.com/token",
-				ProxyService:            "upstream",
-				AllowInsecureProxyRealm: true,
-			},
-		)
-		if !errors.Is(err, errInvalidTokenProxyForm) {
-			t.Fatalf("expected invalid form error, got %v", err)
-		}
-	})
-
-	t.Run("returns request construction errors", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodGet, "/token", nil)
-		request.Method = "bad\nmethod"
-
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			httptest.NewRecorder(),
-			request,
-			&config.BearerConfig{
-				ProxyRealm:              "http://example.com/token",
-				ProxyService:            "upstream",
-				AllowInsecureProxyRealm: true,
-			},
-		)
-		if err == nil {
-			t.Fatal("expected request construction error")
-		}
-	})
-
-	t.Run("returns upstream request errors", func(t *testing.T) {
-		t.Parallel()
-
-		proxyServer := httptest.NewServer(http.NotFoundHandler())
-		proxyRealm := proxyServer.URL + "/token"
-		proxyServer.Close()
-
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			httptest.NewRecorder(),
-			httptest.NewRequest(http.MethodGet, "/token", nil),
-			&config.BearerConfig{ProxyRealm: proxyRealm, ProxyService: "upstream", AllowInsecureProxyRealm: true},
-		)
-		if err == nil {
-			t.Fatal("expected upstream request error")
-		}
-	})
-
-	t.Run("relays redirect responses", func(t *testing.T) {
-		t.Parallel()
-
-		proxyServer := httptest.NewServer(http.HandlerFunc(func(response http.ResponseWriter, request *http.Request) {
-			http.Redirect(response, request, "/next", http.StatusFound)
-		}))
-		defer proxyServer.Close()
-
-		response := httptest.NewRecorder()
-		err := routeHandler.proxyOIDCBearerTokenExchange(
-			response,
-			httptest.NewRequest(http.MethodGet, "/token", nil),
-			&config.BearerConfig{
-				ProxyRealm:              proxyServer.URL + "/token",
-				ProxyService:            "upstream",
-				AllowInsecureProxyRealm: true,
-			},
-		)
-		if err != nil {
-			t.Fatalf("unexpected proxy error: %v", err)
-		}
-
-		if response.Code != http.StatusFound {
-			t.Fatalf("expected status %d, got %d", http.StatusFound, response.Code)
-		}
-
-		if response.Header().Get("Location") != "/next" {
-			t.Fatalf("expected redirect location to be relayed, got %q", response.Header().Get("Location"))
-		}
-	})
-}
-
-func TestOIDCBearerTokenExchangeProxyError(t *testing.T) {
-	t.Parallel()
-
-	conf := config.New()
-	conf.HTTP.Auth = &config.AuthConfig{
-		Bearer: &config.BearerConfig{
-			Realm:        "zot",
-			Service:      "zot",
-			ProxyRealm:   "/token",
-			ProxyService: "upstream",
-		},
-	}
-
-	routeHandler := &RouteHandler{
-		c: &Controller{
-			Config: conf,
-			Log:    log.NewTestLogger(),
-		},
-	}
-
-	request := httptest.NewRequest(http.MethodGet, "/zot/auth/token", nil)
-	request.SetBasicAuth("user", "not-an-oidc-token")
-
-	response := httptest.NewRecorder()
-	routeHandler.OIDCBearerTokenExchange(&OIDCBearerAuthorizer{})(response, request)
-
-	if response.Code != http.StatusBadGateway {
-		t.Fatalf("expected status %d, got %d", http.StatusBadGateway, response.Code)
-	}
-}
-
-func TestOIDCBearerTokenExchangeRejectsUnsupportedMethod(t *testing.T) {
-	t.Parallel()
-
-	conf := config.New()
-	conf.HTTP.Auth = &config.AuthConfig{Bearer: &config.BearerConfig{Realm: "zot"}}
-
-	routeHandler := &RouteHandler{
-		c: &Controller{
-			Config: conf,
-			Log:    log.NewTestLogger(),
-		},
-	}
-
-	request := httptest.NewRequest(http.MethodPut, "/zot/auth/token", nil)
-	response := httptest.NewRecorder()
-
-	routeHandler.OIDCBearerTokenExchange(&OIDCBearerAuthorizer{})(response, request)
-
-	if response.Code != http.StatusMethodNotAllowed {
-		t.Fatalf("expected status %d, got %d", http.StatusMethodNotAllowed, response.Code)
-	}
-}
-
-func TestOIDCBearerTokenExchangeRejectsMultipleAuthorizationHeaders(t *testing.T) {
-	t.Parallel()
-
-	conf := config.New()
-	conf.HTTP.Auth = &config.AuthConfig{Bearer: &config.BearerConfig{Realm: "zot"}}
-
-	routeHandler := &RouteHandler{
-		c: &Controller{
-			Config: conf,
-			Log:    log.NewTestLogger(),
-		},
-	}
-
-	request := httptest.NewRequest(http.MethodGet, "/zot/auth/token", nil)
-	request.Header.Add("Authorization", "Basic dXNlcjpwYXNz")
-	request.Header.Add("Authorization", "Bearer token")
-	response := httptest.NewRecorder()
-
-	routeHandler.OIDCBearerTokenExchange(&OIDCBearerAuthorizer{})(response, request)
-
-	if response.Code != http.StatusUnauthorized {
-		t.Fatalf("expected status %d, got %d", http.StatusUnauthorized, response.Code)
-	}
-}
-
-func TestOIDCBearerTokenExchangeRequestParseErrors(t *testing.T) {
-	t.Parallel()
-
-	conf := config.New()
-	conf.HTTP.Auth = &config.AuthConfig{Bearer: &config.BearerConfig{Realm: "zot"}}
-
-	routeHandler := &RouteHandler{
-		c: &Controller{
-			Config: conf,
-			Log:    log.NewTestLogger(),
-		},
-	}
-
-	t.Run("rejects invalid form body", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, constants.TokenPath, strings.NewReader("password=%zz"))
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-		response := httptest.NewRecorder()
-
-		routeHandler.OIDCBearerTokenExchange(&OIDCBearerAuthorizer{})(response, request)
-
-		if response.Code != http.StatusBadRequest {
-			t.Fatalf("expected status %d, got %d", http.StatusBadRequest, response.Code)
-		}
-	})
-
-	t.Run("rejects unreadable form body", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, constants.TokenPath, nil)
-		request.Body = errTokenProxyReadCloser{}
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-		response := httptest.NewRecorder()
-
-		routeHandler.OIDCBearerTokenExchange(&OIDCBearerAuthorizer{})(response, request)
-
-		if response.Code != http.StatusBadRequest {
-			t.Fatalf("expected status %d, got %d", http.StatusBadRequest, response.Code)
-		}
-	})
-
-	t.Run("rejects oversized form body", func(t *testing.T) {
-		t.Parallel()
-
-		request := httptest.NewRequest(http.MethodPost, constants.TokenPath,
-			strings.NewReader(strings.Repeat("a", constants.MaxTokenRequestBodySize+1)))
-		request.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-		response := httptest.NewRecorder()
-
-		routeHandler.OIDCBearerTokenExchange(&OIDCBearerAuthorizer{})(response, request)
-
-		if response.Code != http.StatusRequestEntityTooLarge {
-			t.Fatalf("expected status %d, got %d", http.StatusRequestEntityTooLarge, response.Code)
-		}
-	})
-}
-
-func TestOIDCBearerTokenExchangeOwnedTokenWithoutAuthorizer(t *testing.T) {
-	t.Parallel()
-
-	const issuer = "https://issuer.example.com"
-	const audience = "zot"
-
-	conf := config.New()
-	conf.HTTP.Auth = &config.AuthConfig{Bearer: &config.BearerConfig{
-		Realm: "zot",
-		OIDC:  []config.BearerOIDCConfig{{Issuer: issuer, Audiences: []string{audience}}},
-	}}
-
-	routeHandler := &RouteHandler{
-		c: &Controller{
-			Config: conf,
-			Log:    log.NewTestLogger(),
-		},
-	}
-
-	request := httptest.NewRequest(http.MethodGet, constants.TokenPath, nil)
-	request.SetBasicAuth("<token>", testJWT(map[string]any{"iss": issuer, "aud": audience}))
-	response := httptest.NewRecorder()
-
-	routeHandler.OIDCBearerTokenExchange(nil)(response, request)
-
-	if response.Code != http.StatusUnauthorized {
-		t.Fatalf("expected status %d, got %d", http.StatusUnauthorized, response.Code)
 	}
 }
